recipes/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
autopep8
isort